feat: add yup schema, remove defaultValues, move file to separate folder, header to separate component
