+++ src/modules/Loads/Loads/components/AddLoad/AddLoad.module.scss
@@ -1,0 +1,56 @@
+@import '../../../../../styles/variables.scss';
+
+.container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+  flex-direction: column;
+  width: 70vw;
+  margin: 0 auto;
+  p,
+  label {
+    font-size: 0.8rem;
+  }
+
+  form {
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    gap: 30px;
+    background-color: rgb(255, 255, 255);
+    width: 100%;
+    padding: 1rem;
+    .inputs-container {
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      gap: 30px;
+    }
+
+    .payment {
+      display: flex;
+    }
+    .title {
+      margin-top: 0;
+    }
+  }
+  .vehicle-inputs {
+    padding: 0.5rem;
+    background-color: $secondary-color;
+  }
+}
+.loading-inputs,
+.unloading-inputs,
+.payment-inputs,
+.vehicle-inputs {
+  padding: 0.5rem;
+  background-color: $secondary-color;
+  border-radius: 0.5rem;
+}
+.submit {
+  background-color: $primary-button-color;
+  color: white;
+  border: none;
+  padding: 0.2rem;
+  border-radius: 0.2rem;
+}
